Crate elrond

Elrond reads Elf files

This is an embryonic Elf Parser written in NOM. It is not feature complete. It requires a lot more testing this is just a cleaned up version I can throw on crates.io to start testing with.

There are no benchmarks.

There are no cool features. It doesn't support dynamic objects (yet).

It isn't super ergonomic. I more-or-less read BSD source code and a pdf to produce this. I'm not even sure it works.

Macros

build_enum

generate_reader

parse_test

std_val

Structs

Elf
Holds the ELF Data

ElfHeader
Full Elf Header

Elf_Half

Elf_SHalf

Elf_SWord

Elf_Sxword

Elf_Word

Elf_XWord

ProgramHeaderTable

Section

Enums

ABI

Arch

Elf_Addr

Elf_Off

Elf_VarWord

IResult
Holds the result of parsing functions

Kind

RamFlags

SectionType

SegFlag